RAG-Wikipedia/requirements.txt
pyngrok
streamlit
einops
sentencepiece
trl
peft
accelerate
bitsandbytes
python-dotenv
openai
llama-index
nltk
sentence_transformers
wikipedia
llama-index-llms-huggingface
llama-index-embeddings-langchain
langchain
langchain-community
